backend/services/python-ai-service/src/requirements.txt
# Core AI/ML Libraries
torch>=2.0.0
transformers>=4.30.0
openai>=1.0.0
anthropic>=0.3.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
nltk>=3.8.0
spacy>=3.6.0

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0

# Database & Cache
sqlalchemy>=2.0.0
asyncpg>=0.28.0
redis>=4.5.0
alembic>=1.11.0

# Communication
httpx>=0.24.0
websockets>=11.0.0
aiofiles>=23.0.0

# Monitoring & Logging
prometheus-client>=0.17.0
opentelemetry-api>=1.17.0
opentelemetry-sdk>=1.17.0
opentelemetry-instrumentation-fastapi>=0.41b0
structlog>=23.0.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
typer>=0.9.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Image Processing
pillow>=10.0.0
opencv-python>=4.8.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0

# Text Processing
textstat>=0.7.0
vaderSentiment>=3.3.2
gensim>=4.3.0

# Memory & Storage
qdrant-client>=1.3.0
chromadb>=0.4.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
mypy>=1.4.0
